Moved the test class to appropriate package
